Commiting small changes with delete and update

+++ BlazorWASM-AppLogic/Application/LogicInterfaces/IPostLogic.cs
@@ -8,7 +8,7 @@ namespace Application.LogicInterfaces;
 
 public interface IPostLogic
 {
-    Task<int> CreateAsync(PostCreationDto postCreationDto);
+    Task<long> CreateAsync(PostCreationDto postCreationDto);
     Task DeleteAsync(long post_id, int user_id);
     Task<int> UpdateAsync(UpdatePostDto dto, int user_id);
